Desktop Login

+++ Library/CloudVOffice.Services/Projects/IProjectTypeService.cs
@@ -15,9 +15,9 @@ namespace CloudVOffice.Services.Projects
     {
         public MennsageEnum ProjectTypeCreate(ProjectTypeDTO projecttypeDTO);
         public List<ProjectType> GetProjectTypes();
-        public ProjectType GetProjectTypeByProjectTypeId(Int64 projecttypeId);
+        public ProjectType GetProjectTypeByProjectTypeId(int projecttypeId);
         public ProjectType GetProjectTypeByProjectName(string projectName);
         public MennsageEnum ProjectTypeUpdate(ProjectTypeDTO projecttypeDTO);
-        public MennsageEnum ProjectTypeDelete(Int64 projecttypeId, Int64 DeletedBy);
+        public MennsageEnum ProjectTypeDelete(int projecttypeId, Int64 DeletedBy);
     }
 }
